moving redis cache to examples folder
