+++ examples/redis_cache/redis_cache.py
